bigquery/types/SchemaUtil.scala Scala 3
